refactor(economy): phase-7 kitchen-sink splits, support densify, size ratchets
Split treasury_bid_budget.dart (spend/available/caps siblings),
town_manufacturing_bonus.dart (math/delivered-raw siblings),
first_right_credits.dart (embassy-kickback sibling), and
tile_extraction_pipeline.dart (context sibling) into single-concern
libraries behind stable barrels; densify colonizethis_economy_test_support
to 7,492 physical LOC; add repo.economy_source_file_size gate (<=310) and
ratchet repo.economy_test_support_loc 7999 -> 7495.

Refs #4049
