+++ packages/colonizethis_economy/lib/src/economy/town_manufacturing_bonus.dart
@@ -4,136 +4,23 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 
 import 'commodity_totals.dart';
-import 'game_lookup_helpers.dart';
-import 'town_connected_province_extraction.dart';
+import 'town_manufacturing_bonus_math.dart';
+import 'town_manufacturing_delivered_raw.dart';
 
-/// Town manufacturing clustering bonus (Refs #3872).
+export 'town_manufacturing_bonus_math.dart';
+export 'town_manufacturing_delivered_raw.dart';
+
+/// Town manufacturing clustering bonus — game rollup, overlay preview, and
+/// auto-offers bridge (Refs #3872; phase-7 split Refs #4049).
 /// SPEC/game/extraction-and-improvements.md § Town manufacturing bonus.
-
-const int kTownManufacturingBonusDivisor = 4;
-
-/// Multiplier for qualifying [townDevelopmentLevel] values (2 → 1, 4 → 2).
-int townManufacturingBonusMultiplier(int townDevelopmentLevel) {
-  return switch (townDevelopmentLevel) {
-    2 => 1,
-    4 => 2,
-    _ => 0,
-  };
-}
-
-bool isTownManufacturingRecipeEligible(ProductionRecipe recipe) {
-  for (final inputId in recipe.inputQuantities.keys) {
-    final commodity = CommodityCatalog.byId[inputId];
-    if (commodity == null ||
-        commodity.category != CommodityCategory.rawMaterial) {
-      return false;
-    }
-  }
-  return true;
-}
-
-/// Per-province manufactured bonus quantities from town-connected delivered
-/// raw extraction.
-Map<CommodityId, int> computeTownManufacturingBonusForProvince({
-  required int townDevelopmentLevel,
-  required Map<CommodityId, int> townConnectedDeliveredRawByCommodity,
-  required Map<String, bool>? techUnlocked,
-}) {
-  final multiplier = townManufacturingBonusMultiplier(townDevelopmentLevel);
-  if (multiplier <= 0 || townConnectedDeliveredRawByCommodity.isEmpty) {
-    return const {};
-  }
-  final bonus = <CommodityId, int>{};
-  for (final recipe in ProductionRecipesCatalog.all) {
-    if (!isTownManufacturingRecipeEligible(recipe)) continue;
-    if (!ProductionRecipesCatalog.isRecipeAvailableForPlayer(
-      recipe,
-      techUnlocked,
-    )) {
-      continue;
-    }
-    var limiting = -1;
-    for (final entry in recipe.inputQuantities.entries) {
-      final rawQty = townConnectedDeliveredRawByCommodity[entry.key] ?? 0;
-      limiting = limiting < 0
-          ? rawQty
-          : (rawQty < limiting ? rawQty : limiting);
-    }
-    if (limiting <= 0) continue;
-    final outputQty =
-        (limiting ~/ kTownManufacturingBonusDivisor) *
-        multiplier *
-        recipe.outputQuantity;
-    if (outputQty <= 0) continue;
-    addUnits(bonus, recipe.outputCommodityId, outputQty);
-  }
-  return bonus;
-}
-
-/// Per-province town-connected **delivered** raw extraction keyed by province id.
-typedef ProvinceDeliveredRawExtraction = Map<String, Map<CommodityId, int>>;
+///
+/// Recipe eligibility / per-province math lives in
+/// `town_manufacturing_bonus_math.dart`; the town-connected delivered-raw
+/// province walk lives in `town_manufacturing_delivered_raw.dart`. Both are
+/// re-exported here so callers keep one stable import surface.
 
 /// Per-faction aggregated manufactured bonus keyed by faction id.
 typedef FactionManufacturingBonus = Map<String, Map<CommodityId, int>>;
-
-ProvinceDeliveredRawExtraction computeTownConnectedDeliveredRawByProvince({
-  required Game game,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required Map<String, ConnectivityResult> gpConnectivityByPlayerId,
-  required Map<String, ConnectivityResult> nonGpConnectivityByFactionId,
-  required Map<String, Set<String>> townConnectedByProvinceId,
-}) {
-  final provincesByFullId = buildProvinceIndex(game);
-  final portTileKeys = collectPortTileKeys(game);
-  final landByProvince = <String, Map<CommodityId, int>>{};
-  final overseasByPlayerProvince =
-      <String, Map<String, Map<CommodityId, int>>>{};
-  final ctx = TownConnectedProvinceExtractionContext(
-    game: game,
-    tileMapByRegion: tileMapByRegion,
-    provincesByFullId: provincesByFullId,
-    portTileKeys: portTileKeys,
-    landByProvince: landByProvince,
-    overseasByPlayerProvince: overseasByPlayerProvince,
-  );
-
-  for (final province in allProvinces(game.worldState)) {
-    final ownerId = province.ownerId;
-    if (ownerId == null || ownerId.isEmpty) continue;
-    final townConnected = townConnectedByProvinceId[province.id];
-    if (townConnected == null || townConnected.isEmpty) continue;
-
-    final player = game.playerById(ownerId);
-    if (player != null) {
-      accumulateGpProvinceExtraction(
-        ctx: ctx,
-        player: player,
-        province: province,
-        townConnected: townConnected,
-        connectivity: gpConnectivityByPlayerId[player.id],
-      );
-      continue;
-    }
-
-    final connectivity = nonGpConnectivityByFactionId[ownerId];
-    if (connectivity == null) continue;
-    accumulateNonGpProvinceExtraction(
-      ctx: ctx,
-      province: province,
-      townConnected: townConnected,
-      connectivity: connectivity,
-      ownerId: ownerId,
-    );
-  }
-
-  applyOverseasCargoDeliveryToProvinces(
-    game: game,
-    overseasByPlayerProvince: overseasByPlayerProvince,
-    landByProvince: landByProvince,
-  );
-
-  return landByProvince;
-}
 
 /// Per-province manufactured bonus keyed by full province id.
 typedef ProvinceManufacturingBonus = Map<String, Map<CommodityId, int>>;
