+++ packages/colonizethis_economy/lib/src/economy/world_market/treasury_bid_budget.dart
@@ -1,286 +1,25 @@
-/// Treasury budget helpers for the world-market bid clamp on the Trade
-/// Screen Market tab.
+/// Treasury budget surface for the world-market bid clamp on the Trade
+/// Screen Market tab (Refs #3093; phase-7 split Refs #4049).
 ///
 /// SPEC/game/world-market.md § Treasury budget for bids,
 /// SPEC/ui/trade-screen.md § Market tab — treasury bid cap.
 ///
-/// The helpers are **pure**: deterministic for fixed inputs, silent (no
-/// logger calls), and safe to call from order-submission, AI suggestion,
-/// resolver-prep, and per-frame UI paths under the 15-second
-/// turn-resolution budget (`.cursor/rules/colonizethis-turn-resolution-budget.mdc`).
+/// Barrel over the three single-concern sibling libraries so every
+/// existing call site (validator, suggester, deal-matcher session, UI,
+/// AI planners) keeps one stable import:
 ///
-/// Scope (#3093 — treasury bid budget slices):
+/// * `treasury_bid_spend.dart` — effective price lookup and per-order /
+///   staged / carry-forward bid **spend** totals.
+/// * `treasury_bid_available.dart` — the player's **available** bid
+///   budget with the projected non-bid deficit subtracted.
+/// * `treasury_bid_caps.dart` — bid quantity **caps** and the
+///   fill-time treasury **decrement** shared by matcher, suggester,
+///   and validator.
 ///
-/// * `effectiveMarketPriceForCommodityId` and `stagedBidTotalSpendByPlayer`
-///   compose the per-player running bid-spend total from
-///   `currentOrdersProvider`.
-/// * `treasuryAvailableForBidsByPlayer` returns the player's bid budget.
-///   When the caller supplies `projectedNonBidTreasuryDelta` (the signed
-///   treasury change from the player's **non-bid** staged orders this turn
-///   — i.e. `projectOrderEffects(orders).treasuryDelta` plus the player's
-///   own running bid spend so the bid contribution is netted back out)
-///   the helper subtracts the projected **deficit** from raw treasury
-///   (a non-positive delta becomes a positive deficit; a positive delta
-///   is ignored so net income from non-bid orders never raises the
-///   budget). With the default `projectedNonBidTreasuryDelta == 0` the
-///   helper falls back to the legacy "raw treasury" contract for
-///   callers that don't run a projection.
-///
-/// Validator-side enforcement lives in `trade_order_validator.dart` (rule 5).
+/// Validator-side enforcement lives in `trade_order_validator.dart`
+/// (rule 5).
 library;
 
-import 'package:colonizethis_data/colonizethis_data.dart' as data;
-import 'package:colonizethis_models/colonizethis_models.dart';
-
-import 'package:colonizethis_world/colonizethis_world.dart'
-    show GamePlayerLookup;
-
-import 'trade_order_admission.dart' show isWorldMarketTradeableCommodity;
-
-/// Effective per-unit market price used by bid budget logic.
-///
-/// Reads the integer treasury-unit price stored on
-/// [WorldMarketState.prices] (per `SPEC/game/world-market.md` § Price
-/// discovery), falling back to
-/// `resourceRules.defaultMarketPriceForCommodityId(commodityId)` when
-/// the market state lacks an entry for [commodityId] (typically on
-/// first-turn games before price discovery runs). The catalog default
-/// now covers every tradeable commodity — raw resources (per the
-/// `Resource` enum default-price map) and manufactured commodities
-/// (per `SPEC/game/commodity-catalog.md` § *Manufactured base prices*) —
-/// so `null` is reserved for the defensive case where a future
-/// tradeable commodity ships without a catalog default, or when
-/// [commodityId] is in [data.richesCommodityIds] (those do not trade on
-/// the world market per SPEC § Tradeable commodities).
-///
-/// Callers should treat a `null` return as "no spend contribution"
-/// (`0`) so bid budget math remains defensive against missing-data
-/// states — UI consumers cannot stage a bid whose price is unknown
-/// because the row's price text reads as the em-dash in that case.
-int? effectiveMarketPriceForCommodityId({
-  required CommodityId commodityId,
-  required WorldMarketState worldMarket,
-  required data.ResourceRules resourceRules,
-}) {
-  if (!isWorldMarketTradeableCommodity(commodityId)) return null;
-  final int? stored = worldMarket.prices[commodityId];
-  if (stored != null && stored >= 0) return stored;
-  return resourceRules.defaultMarketPriceForCommodityId(commodityId);
-}
-
-/// Treasury spend for a single bid order: `quantity × effectiveMarketPrice`.
-///
-/// Returns `0` for offers, non-positive quantities, or when no effective
-/// price exists (`null` from [effectiveMarketPriceForCommodityId]).
-int bidTreasurySpendForOrder({
-  required TradeOrder order,
-  required WorldMarketState worldMarket,
-  required data.ResourceRules resourceRules,
-}) {
-  if (order.type != TradeOrderType.bid || order.quantity <= 0) return 0;
-  final int? price = effectiveMarketPriceForCommodityId(
-    commodityId: order.commodityId,
-    worldMarket: worldMarket,
-    resourceRules: resourceRules,
-  );
-  if (price == null) return 0;
-  return order.quantity * price;
-}
-
-/// Sum of `quantity × effectiveMarketPrice` across every staged
-/// `TradeOrderType.bid` for [playerId] in [orders].
-///
-/// Offers do not contribute to bid spend. Non-positive quantities and
-/// bids on commodities with no effective price (missing in both
-/// `worldMarket.prices` and the catalog default) are skipped — the UI
-/// clamp would not let those bids land in the first place, but the
-/// helper stays defensive in case AI or save state injects an
-/// out-of-band order.
-int stagedBidTotalSpendByPlayer({
-  required Orders orders,
-  required String playerId,
-  required Game game,
-  required data.ResourceRules resourceRules,
-}) {
-  return _sumBidSpend(
-    orders: orders.tradeOrdersByPlayerId[playerId],
-    game: game,
-    resourceRules: resourceRules,
-  );
-}
-
-/// Sum of `quantity × effectiveMarketPrice` over the bid orders in [orders].
-///
-/// Shared core for [stagedBidTotalSpendByPlayer] (staged orders) and
-/// [carryForwardBidNotionalByPlayer] (carry-forward bids): the only
-/// difference between the two callers is the source iterable. Offers,
-/// non-positive quantities, and bids on commodities with no effective
-/// price are skipped so the total stays defensive against out-of-band
-/// orders injected by AI or save state. Returns `0` for a `null` or
-/// empty [orders].
-int _sumBidSpend({
-  required List<TradeOrder>? orders,
-  required Game game,
-  required data.ResourceRules resourceRules,
-}) {
-  if (orders == null || orders.isEmpty) return 0;
-  int total = 0;
-  for (final TradeOrder o in orders) {
-    total += bidTreasurySpendForOrder(
-      order: o,
-      worldMarket: game.worldMarketState,
-      resourceRules: resourceRules,
-    );
-  }
-  return total;
-}
-
-/// Sum of `quantity × effectiveMarketPrice` across every
-/// **carry-forward** `TradeOrderType.bid` for [playerId] held on
-/// `game.worldMarketState.carryForwardBidsByFactionId[playerId]`.
-///
-/// Carry-forward bids are still "live" from the prior turn and will
-/// be re-matched at phase 13, so the matcher (#3115) clamps the
-/// current-turn buyer treasury against them. AI / planner callers that
-/// pre-size new bids must therefore subtract this notional from the
-/// raw `player.treasury` before sizing any new bid.
-///
-/// Non-positive quantities and bids on commodities with no effective
-/// price (manufactured commodities until in-game price discovery
-/// seeds a price) contribute `0` to the total, matching the validator-
-/// side defensive skip in [stagedBidTotalSpendByPlayer].
-///
-/// Returns `0` when [playerId] has no carry-forward bids.
-int carryForwardBidNotionalByPlayer({
-  required Game game,
-  required String playerId,
-  required data.ResourceRules resourceRules,
-}) {
-  return _sumBidSpend(
-    orders: game.worldMarketState.carryForwardBidsByFactionId[playerId],
-    game: game,
-    resourceRules: resourceRules,
-  );
-}
-
-/// Treasury budget [playerId] may commit to bids this turn.
-///
-/// Returns `max(0, treasury − pendingNonBidDeficit)`, where
-/// `pendingNonBidDeficit = max(0, −projectedNonBidTreasuryDelta)`. The
-/// helper therefore stays **conservative**:
-///
-/// * A negative `projectedNonBidTreasuryDelta` (the player's non-bid
-///   staged orders — build / recruit / civilian work / subsidies — would
-///   net-debit treasury this turn) reduces the bid budget by exactly
-///   that deficit.
-/// * A non-negative `projectedNonBidTreasuryDelta` (net income or
-///   neutral) leaves the bid budget at the raw `treasury` value; net
-///   non-bid income never raises the budget so the clamp never lets the
-///   player commit treasury they only project to earn.
-///
-/// `projectedNonBidTreasuryDelta` defaults to `0`, which preserves the
-/// legacy "raw treasury" contract for callers that do not run a
-/// projection (e.g. Widgetbook stories, isolated widget tests, and AI
-/// suggestion paths that have not yet wired projection).
-///
-/// Callers (UI + validator) factor in the player's own staged **other**
-/// bid spend by composing this helper with [stagedBidTotalSpendByPlayer],
-/// so the effective treasury budget for the row currently being clamped
-/// is:
-///
-///     treasuryAvailableForBidsByPlayer(...) - (
-///        stagedBidTotalSpendByPlayer(...) - rowPriorBidSpend
-///     )
-///
-/// The Trade Screen Market tab computes
-/// `projectedNonBidTreasuryDelta` as
-/// `treasurySummaryProvider.projectedDelta + stagedBidTotalSpendByPlayer(...)`
-/// (`projectedDelta` is the signed `projectOrderEffects(orders)` net
-/// treasury change including bids; adding the player's running bid spend
-/// back reconstructs the non-bid contribution) per
-/// `SPEC/ui/trade-screen.md` § Market tab — treasury bid cap. When the
-/// projection is unavailable
-/// (`treasurySummaryProvider.projectedDelta == null`) the UI passes `0`
-/// and the helper falls back to raw treasury.
-///
-/// Returns `0` when [playerId] does not resolve to a player.
-int treasuryAvailableForBidsByPlayer({
-  required Game game,
-  required String playerId,
-  int projectedNonBidTreasuryDelta = 0,
-}) {
-  final player = game.playerById(playerId);
-  if (player == null) return 0;
-  final int treasury = player.treasury;
-  if (treasury <= 0) return 0;
-  final int pendingDeficit = projectedNonBidTreasuryDelta < 0
-      ? -projectedNonBidTreasuryDelta
-      : 0;
-  final int budget = treasury - pendingDeficit;
-  return budget < 0 ? 0 : budget;
-}
-
-/// Caps [bidQuantity] by [remainingCargoBudget] and affordable treasury spend.
-///
-/// When [unitPrice] is null or non-positive, only the cargo cap applies.
-/// Used by [TradeOrderSuggester] for per-bid quantity clamping; the
-/// validator enforces the same constraints sequentially per order (rule 5).
-int capBidQuantityForBudgets({
-  required int bidQuantity,
-  required int remainingCargoBudget,
-  required int remainingTreasuryBudget,
-  required int? unitPrice,
-}) {
-  if (bidQuantity <= 0 || remainingCargoBudget <= 0) return 0;
-  var cappedQty = bidQuantity < remainingCargoBudget
-      ? bidQuantity
-      : remainingCargoBudget;
-  if (unitPrice != null && unitPrice > 0) {
-    final int maxAffordable = remainingTreasuryBudget ~/ unitPrice;
-    if (cappedQty > maxAffordable) {
-      cappedQty = maxAffordable;
-    }
-  }
-  return cappedQty <= 0 ? 0 : cappedQty;
-}
-
-/// Maximum bid quantity affordable at [pricePerUnit] given
-/// [remainingTreasuryBudget].
-///
-/// When [pricePerUnit] is non-positive, returns [bidRemaining] so the
-/// deal matcher preserves the missing-price free-fill contract per
-/// `SPEC/program/world-market-resolution.md` § Step C (Refs #3115): the
-/// treasury clamp is skipped and the other three clamps dominate.
-///
-/// Used by [DealMatcher] `_attemptMatch`; [capBidQuantityForBudgets]
-/// applies the same semantics for integer [unitPrice] on the suggester
-/// and validator paths.
-int maxAffordableBidQuantity({
-  required int bidRemaining,
-  required double pricePerUnit,
-  required int remainingTreasuryBudget,
-}) {
-  if (pricePerUnit <= 0.0) return bidRemaining;
-  if (remainingTreasuryBudget <= 0) return 0;
-  final affordable = (remainingTreasuryBudget / pricePerUnit).floor();
-  return affordable < 0 ? 0 : affordable;
-}
-
-/// Decrements [remainingTreasuryByBuyerFactionId] after a successful match.
-///
-/// Skips the decrement on the missing-price defect path
-/// (`pricePerUnit <= 0`) so free-fill behavior is preserved — no treasury
-/// accounting when no notional is owed. The matcher, suggester, and
-/// validator share this module for treasury affordability (Refs #3856).
-void decrementTreasuryForFill({
-  required String buyerFactionId,
-  required int matchQty,
-  required double pricePerUnit,
-  required Map<String, int> remainingTreasuryByBuyerFactionId,
-}) {
-  if (pricePerUnit <= 0.0 || matchQty <= 0) return;
-  final notional = (matchQty * pricePerUnit).round();
-  final treasuryLeft = remainingTreasuryByBuyerFactionId[buyerFactionId] ?? 0;
-  final next = treasuryLeft - notional;
-  remainingTreasuryByBuyerFactionId[buyerFactionId] = next < 0 ? 0 : next;
-}
+export 'treasury_bid_available.dart';
+export 'treasury_bid_caps.dart';
+export 'treasury_bid_spend.dart';
